suppression pour l'admin pour listeFilm et abonnement

+++ cinemvBack/DTO/NoteDTO.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 public class NoteDTO
 {
     public int Id { get; set; }
-    public int Valeur { get; set; }
+    public double Valeur { get; set; }
     public DateTime DateCreation { get; set; }
     public int UtilisateurId { get; set; }
     public string FilmId { get; set; } = null!;
