+++ cinemvBack/Models/Note.cs
@@ -3,7 +3,7 @@ using cinemvBack.Models;
 public class Note
 {
     public int Id { get; set; }
-    public int Valeur { get; set; }
+    public double Valeur { get; set; }
     public DateTime DateCreation { get; set; } //=DateTime.Now;
     public int UtilisateurId { get; set; }
     public Utilisateur Utilisateur { get; set; } = null!;
